requirements.txt
Django>=4.0
djangorestframework>=3.14
Pillow>=10.0
qrcode[pil]>=7.4
cairosvg>=2.7
celery>=5.3
redis>=4.5
django-celery-beat>=2.5
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
pandas>=2.0
openpyxl>=3.1
twilio>=8.0
python-dotenv>=1.0
django-filter>=23.3
reportlab>=4.0  # For PDF generation
whitenoise>=6.5  # For static files
gunicorn>=21.0  # For production
django-celery-results
pycairo
cairosvg

